backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
lxml
python-multipart
httpx
python-docx
python-dotenv